extract private classes and struct to separate file
